Adding a cookie banner
Adding a cookie banner and configuring Google Analytics. Following the
steps in this tutorial: https://www.improvebadcode.com/gatsby-gdpr-cookie-consent/

Introducing two new dependencies to achieve this:

- react-cookie-consent (https://www.npmjs.com/package/react-cookie-consent) to display a cookie banner and allow the user to accept (or ignore)
- gatsby-plugin-gdpr-cookies (https://www.gatsbyjs.com/plugins/gatsby-plugin-gdpr-cookies/) to only create the Google Analytics cookie if the user accepted the consent

The way they work is because react-cookie-consent creates a cookie when
the user accepts the banner, specifying that they did. Then,
gatsby-plugin-gdpr-cookies only creates the Google Analytics cookies if
the cookie created by react-cookie-consent is present and with a value
of true (i.e. "Accepted")